requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
requests
transformers
torch
matplotlib
pvlib
pandas